+++ CHANGELOG.md
@@ -20,6 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   are stored in Secret resources as a `tar.gz`. If the secret would get too big, the backup can be downloaded from
   the operator pod.
 - Default images:
+  * LINSTOR 1.18.0-rc.3
+  * LINSTOR CSI 0.18.0
+  * DRBD 9.1.6
+  * DRBD Reactor 0.5.3
+  * LINSTOR HA Controller 0.3.0
   * CSI Attacher v3.4.0
   * CSI Node Driver Registrar v2.4.0
   * CSI Provisioner v3.1.0